+++ packages/vue/src/components/IonRouterOutlet.ts
@@ -7,29 +7,21 @@ import {
   provide,
   watch,
   shallowRef,
-  InjectionKey
+  InjectionKey,
+  onUnmounted
 } from 'vue';
-import { AnimationBuilder } from '@ionic/core';
-import { useRoute, useRouter } from 'vue-router';
-import { fireLifecycle, generateId, LIFECYCLE_DID_ENTER, LIFECYCLE_DID_LEAVE, LIFECYCLE_WILL_ENTER, LIFECYCLE_WILL_LEAVE } from '../utils';
+import { AnimationBuilder, LIFECYCLE_DID_ENTER, LIFECYCLE_DID_LEAVE, LIFECYCLE_WILL_ENTER, LIFECYCLE_WILL_LEAVE } from '@ionic/core';
+import { matchedRouteKey, useRoute } from 'vue-router';
+import { fireLifecycle, generateId, getConfig } from '../utils';
 
 let viewDepthKey: InjectionKey<0> = Symbol(0);
 export const IonRouterOutlet = defineComponent({
   name: 'IonRouterOutlet',
   setup(_, { attrs }) {
-    const vueRouter = useRouter();
     const route = useRoute();
     const depth = inject(viewDepthKey, 0);
-
-    // TODO types
-    let tabsPrefix: string | undefined;
     const matchedRouteRef: any = computed(() => {
       const matchedRoute = route.matched[depth];
-
-      if (attrs.tabs && !tabsPrefix) {
-          tabsPrefix = route.matched[0].path;
-          viewStacks.addTabsPrefix(tabsPrefix);
-      }
 
       if (matchedRoute && attrs.tabs && route.matched[depth + 1]) {
         return route.matched[route.matched.length - 1];
@@ -39,6 +31,7 @@ export const IonRouterOutlet = defineComponent({
     });
 
     provide(viewDepthKey, depth + 1)
+    provide(matchedRouteKey, matchedRouteRef);
 
     const ionRouterOutlet = ref();
     const id = generateId('ion-router-outlet');
@@ -51,19 +44,45 @@ export const IonRouterOutlet = defineComponent({
 
     let skipTransition = false;
 
-    watch(matchedRouteRef, () => {
-      setupViewItem(matchedRouteRef);
+    // The base url for this router outlet
+    let parentOutletPath: string;
+
+    watch(matchedRouteRef, (currentValue, previousValue) => {
+      /**
+       * We need to make sure that we are not re-rendering
+       * the same view if navigation changes in a sub-outlet.
+       * This is mainly for tabs when outlet 1 renders ion-tabs
+       * and outlet 2 renders the individual tab view. We don't
+       * want outlet 1 creating a new ion-tabs instance every time
+       * we switch tabs.
+       */
+      if (currentValue !== previousValue) {
+        setupViewItem(matchedRouteRef);
+      }
     });
 
     const canStart = () => {
+      const config = getConfig();
+      const swipeEnabled = config && config.get('swipeBackEnabled', ionRouterOutlet.value.mode === 'ios');
+      if (!swipeEnabled) return false;
+
       const stack = viewStacks.getViewStack(id);
-      return stack && stack.length > 1;
+      if (!stack || stack.length <= 1) return false;
+
+      /**
+       * We only want to outlet of the entering view
+       * to respond to this gesture, so check
+       * to make sure the view is in the outlet we want.
+       */
+      const routeInfo = ionRouter.getCurrentRouteInfo();
+      const enteringViewItem = viewStacks.findViewItemByRouteInfo({ pathname: routeInfo.pushedByRoute }, id);
+
+      return !!enteringViewItem;
     }
     const onStart = async () => {
       const routeInfo = ionRouter.getCurrentRouteInfo();
       const { routerAnimation } = routeInfo;
-      const items = viewStacks.getViewStack(id);
-      const enteringViewItem = items[items.length - 2];
+      const enteringViewItem = viewStacks.findViewItemByRouteInfo({ pathname: routeInfo.pushedByRoute }, id);
       const leavingViewItem = viewStacks.findViewItemByRouteInfo(routeInfo, id);
 
       if (leavingViewItem) {
@@ -105,7 +124,23 @@ export const IonRouterOutlet = defineComponent({
     const onEnd = (shouldContinue: boolean) => {
       if (shouldContinue) {
         skipTransition = true;
-        vueRouter.back();
+
+        /**
+         * Use the same logic as clicking
+         * ion-back-button to determine where
+         * to go back to.
+         */
+        ionRouter.handleNavigateBack();
+      } else {
+        /**
+         * In the event that the swipe
+         * gesture was aborted, we should
+         * re-hide the page that was going to enter.
+         */
+        const routeInfo = ionRouter.getCurrentRouteInfo();
+        const enteringViewItem = viewStacks.findViewItemByRouteInfo({ pathname: routeInfo.pushedByRoute }, id);
+        enteringViewItem.ionPageElement.setAttribute('aria-hidden', 'true');
+        enteringViewItem.ionPageElement.classList.add('ion-page-hidden');
       }
     }
 
@@ -164,13 +199,13 @@ export const IonRouterOutlet = defineComponent({
 
       if (enteringViewItem === leavingViewItem) return;
 
-      fireLifecycle(enteringViewItem.vueComponent, LIFECYCLE_WILL_ENTER);
+      fireLifecycle(enteringViewItem.vueComponent, enteringViewItem.vueComponentRef, LIFECYCLE_WILL_ENTER);
 
       if (leavingViewItem) {
         let animationBuilder = routerAnimation;
         const leavingEl = leavingViewItem.ionPageElement;
 
-        fireLifecycle(leavingViewItem.vueComponent, LIFECYCLE_WILL_LEAVE);
+        fireLifecycle(leavingViewItem.vueComponent, leavingViewItem.vueComponentRef, LIFECYCLE_WILL_LEAVE);
 
         /**
         * If we are going back from a page that
@@ -203,7 +238,11 @@ export const IonRouterOutlet = defineComponent({
         leavingEl.classList.add('ion-page-hidden');
         leavingEl.setAttribute('aria-hidden', 'true');
 
-        if (!(routerAction === 'push' && routerDirection === 'forward')) {
+        if (routerAction === 'replace') {
+          leavingViewItem.mount = false;
+          leavingViewItem.ionPageElement = undefined;
+          leavingViewItem.ionRoute = false;
+        } else if (!(routerAction === 'push' && routerDirection === 'forward')) {
           const shouldLeavingViewBeRemoved = routerDirection !== 'none' && leavingViewItem && (enteringViewItem !== leavingViewItem);
           if (shouldLeavingViewBeRemoved) {
             leavingViewItem.mount = false;
@@ -212,7 +251,7 @@ export const IonRouterOutlet = defineComponent({
           }
         }
 
-        fireLifecycle(leavingViewItem.vueComponent, LIFECYCLE_DID_LEAVE);
+        fireLifecycle(leavingViewItem.vueComponent, leavingViewItem.vueComponentRef, LIFECYCLE_DID_LEAVE);
       } else {
         /**
          * If there is no leaving element, just show
@@ -223,22 +262,35 @@ export const IonRouterOutlet = defineComponent({
         requestAnimationFrame(() => enteringEl.classList.remove('ion-page-invisible'));
       }
 
-      fireLifecycle(enteringViewItem.vueComponent, LIFECYCLE_DID_ENTER);
+      fireLifecycle(enteringViewItem.vueComponent, enteringViewItem.vueComponentRef, LIFECYCLE_DID_ENTER);
 
       components.value = viewStacks.getChildrenToRender(id);
     }
 
-    // TODO types
     const setupViewItem = (matchedRouteRef: any) => {
-      if (!matchedRouteRef.value) {
-        return;
+      const firstMatchedRoute = route.matched[0];
+      if (!parentOutletPath) {
+        parentOutletPath = firstMatchedRoute.path;
+      }
+
+      /**
+       * If no matched route, do not do anything in this outlet.
+       * If there is a match, but it the first matched path
+       * is not the root path for this outlet, then this view
+       * change needs to be rendered in a different outlet.
+       * We also add an exception for when the matchedRouteRef is
+       * equal to the first matched route (i.e. the base router outlet).
+       * This logic is mainly to help nested outlets/multi-tab
+       * setups work better.
+       */
+      if (
+        !matchedRouteRef.value ||
+        (matchedRouteRef.value !== firstMatchedRoute && firstMatchedRoute.path !== parentOutletPath)
+      ) {
+          return;
       }
 
       const currentRoute = ionRouter.getCurrentRouteInfo();
-      const hasTabsPrefix = viewStacks.hasTabsPrefix(currentRoute.pathname)
-      const isLastPathTabs = viewStacks.hasTabsPrefix(currentRoute.lastPathname);
-      if (hasTabsPrefix && isLastPathTabs && !attrs.tabs) { return; }
-
       let enteringViewItem = viewStacks.findViewItemByRouteInfo(currentRoute, id);
 
       if (!enteringViewItem) {
@@ -248,11 +300,10 @@ export const IonRouterOutlet = defineComponent({
 
       if (!enteringViewItem.mount) {
         enteringViewItem.mount = true;
-        enteringViewItem.vueComponent.components.IonPage.mounted = function () {
-          viewStacks.registerIonPage(enteringViewItem, this.$refs.ionPage);
+        enteringViewItem.registerCallback = () => {
           handlePageTransition();
-          enteringViewItem.vueComponent.components.IonPage.mounted = undefined;
-        };
+          enteringViewItem.registerCallback = undefined;
+        }
       } else {
         handlePageTransition();
       }
@@ -264,20 +315,86 @@ export const IonRouterOutlet = defineComponent({
       setupViewItem(matchedRouteRef);
     }
 
+    /**
+     * Remove stack data for this outlet
+     * when outlet is destroyed otherwise
+     * we will see cached view data.
+     */
+    onUnmounted(() => viewStacks.clear(id));
+
+    // TODO types
+    const registerIonPage = (viewItem: any, ionPageEl: HTMLElement) => {
+      const oldIonPageEl = viewItem.ionPageElement;
+
+      viewStacks.registerIonPage(viewItem, ionPageEl);
+
+      /**
+       * If there is a registerCallback,
+       * then this component is being registered
+       * as a result of a navigation change.
+       */
+      if (viewItem.registerCallback) {
+        viewItem.registerCallback();
+
+      /**
+       * If there is no registerCallback, then
+       * this component is likely being re-registered
+       * as a result of a hot module replacement.
+       * We need to see if the oldIonPageEl has
+       * .ion-page-invisible. If it does not then we
+       * need to remove it from the new ionPageEl otherwise
+       * the page will be hidden when it is replaced.
+       */
+      } else if (oldIonPageEl && !oldIonPageEl.classList.contains('ion-page-invisible')) {
+        ionPageEl.classList.remove('ion-page-invisible');
+      }
+  };
     return {
       id,
       components,
-      ionRouterOutlet
+      ionRouterOutlet,
+      registerIonPage
     }
   },
   render() {
-    const { components } = this;
+    const { components, registerIonPage } = this;
 
     return h(
       'ion-router-outlet',
       { ref: 'ionRouterOutlet' },
       // TODO types
-      components && components.map((c: any) => h(c.vueComponent, { key: c.pathname }))
+      components && components.map((c: any) => {
+        let props = {
+          ref: c.vueComponentRef,
+          key: c.pathname,
+          isInOutlet: true,
+          registerIonPage: (ionPageEl: HTMLElement) => registerIonPage(c, ionPageEl)
+        }
+
+        /**
+         * IonRouterOutlet does not support named outlets.
+         */
+        if (c.matchedRoute?.props?.default) {
+          const matchedRoute = c.matchedRoute;
+          const routePropsOption = matchedRoute.props.default;
+          const routeProps = routePropsOption
+            ? routePropsOption === true
+              ? c.params
+              : typeof routePropsOption === 'function'
+              ? routePropsOption(matchedRoute)
+              : routePropsOption
+            : null
+
+          props = {
+            ...props,
+            ...routeProps
+          }
+        }
+        return h(
+          c.vueComponent,
+          props
+        );
+      })
     )
   }
 });