+++ packages/vue/src/index.ts
@@ -12,6 +12,8 @@ export { IonTabs } from './components/IonTabs';
 export { IonTabBar } from './components/IonTabBar';
 export { IonNav } from './components/IonNav';
 export { IonIcon } from './components/IonIcon';
+export { IonApp } from './components/IonApp';
+
 export * from './components/Overlays';
 
 export { IonKeyboardRef, IonRouter, useBackButton, useIonRouter, useKeyboard } from './hooks';
@@ -20,6 +22,8 @@ export {
   modalController,
   popoverController
 } from './controllers';
+
+export * from './globalExtensions';
 
 export {
   // Overlay Controllers