+++ packages/vue/src/proxies.ts
@@ -3,12 +3,9 @@
 /* auto-generated vue proxies */
 import { defineContainer } from './vue-component-lib/utils';
 
-import { JSX } from '@ionic/core';
+import type { JSX } from '@ionic/core';
 
 
-
-
-export const IonApp = /*@__PURE__*/ defineContainer<JSX.IonApp>('ion-app');
 
 
 export const IonAvatar = /*@__PURE__*/ defineContainer<JSX.IonAvatar>('ion-avatar');
@@ -112,7 +109,8 @@ export const IonCheckbox = /*@__PURE__*/ defineContainer<JSX.IonCheckbox>('ion-c
 
 export const IonChip = /*@__PURE__*/ defineContainer<JSX.IonChip>('ion-chip', [
   'color',
-  'outline'
+  'outline',
+  'disabled'
 ]);
 
 
@@ -576,6 +574,7 @@ export const IonSegment = /*@__PURE__*/ defineContainer<JSX.IonSegment>('ion-seg
   'color',
   'disabled',
   'scrollable',
+  'swipeGesture',
   'value',
   'ionChange',
   'ionSelect',