+++ packages/vue/src/utils.ts
@@ -1,7 +1,7 @@
-export const LIFECYCLE_WILL_ENTER = 'ionViewWillEnter';
-export const LIFECYCLE_DID_ENTER = 'ionViewDidEnter';
-export const LIFECYCLE_WILL_LEAVE = 'ionViewWillLeave';
-export const LIFECYCLE_DID_LEAVE = 'ionViewDidLeave';
+import { Ref, ComponentPublicInstance } from 'vue';
+import { Config as CoreConfig, LIFECYCLE_DID_ENTER, LIFECYCLE_DID_LEAVE, LIFECYCLE_WILL_ENTER, LIFECYCLE_WILL_LEAVE } from '@ionic/core';
+
+type LIFECYCLE_EVENTS = typeof LIFECYCLE_WILL_ENTER | typeof LIFECYCLE_DID_ENTER | typeof LIFECYCLE_WILL_LEAVE | typeof LIFECYCLE_DID_LEAVE;
 
 const ids: { [k: string]: number } = { main: 0 };
 
@@ -12,8 +12,23 @@ export const generateId = (type = 'main') => {
 };
 
 // TODO types
-export const fireLifecycle = (vueComponent: any, lifecycle: string) => {
-  if (vueComponent && vueComponent.methods && vueComponent.methods[lifecycle]) {
-    vueComponent.methods[lifecycle]();
+export const fireLifecycle = (vueComponent: any, vueInstance: Ref<ComponentPublicInstance>, lifecycle: LIFECYCLE_EVENTS) => {
+  if (vueComponent?.[lifecycle]) {
+    vueComponent[lifecycle].bind(vueInstance?.value)();
+  }
+
+  const instance = vueInstance?.value as any;
+  if (instance?.[lifecycle]) {
+    instance[lifecycle]();
   }
 }
+
+export const getConfig = (): CoreConfig | null => {
+  if (typeof (window as any) !== 'undefined') {
+    const Ionic = (window as any).Ionic;
+    if (Ionic && Ionic.config) {
+      return Ionic.config;
+    }
+  }
+  return null;
+};