+++ packages/vue/test-app/src/router/index.ts
@@ -1,11 +1,21 @@
 import { createRouter, createWebHistory } from '@ionic/vue-router';
 import { RouteRecordRaw } from 'vue-router';
-import Home from '../views/Home.vue'
+import Home from '@/views/Home.vue'
+import { DelayGuard } from '@/guards/Delay';
 
 const routes: Array<RouteRecordRaw> = [
   {
     path: '/',
     component: Home
+  },
+  {
+    path: '/delayed-redirect',
+    beforeEnter: DelayGuard,
+    component: Home
+  },
+  {
+    path: '/lifecycle',
+    component: () => import('@/views/Lifecycle.vue')
   },
   {
     path: '/overlays',
@@ -27,13 +37,24 @@ const routes: Array<RouteRecordRaw> = [
     component: () => import('@/views/DefaultHref.vue')
   },
   {
-    path: '/navigation',
-    name: 'Navigation',
-    component: () => import('@/views/Navigation.vue')
+    path: '/routing',
+    component: () => import('@/views/Routing.vue')
   },
   {
-    path: '/navigation/child',
-    component: () => import('@/views/NavigationChild.vue')
+    path: '/routing/child',
+    component: () => import('@/views/RoutingChild.vue')
+  },
+  {
+    path: '/routing/:id',
+    component: () => import('@/views/RoutingParameter.vue')
+  },
+  {
+    path: '/routing/:id/view',
+    component: () => import('@/views/RoutingParameterView.vue')
+  },
+  {
+    path: '/navigation',
+    component: () => import('@/views/Navigation.vue')
   },
   {
     path: '/nested',
@@ -78,7 +99,32 @@ const routes: Array<RouteRecordRaw> = [
       },
       {
         path: 'tab3',
+        beforeEnter: (to, from, next) => {
+          next({ path: '/tabs/tab1' });
+        },
         component: () => import('@/views/Tab3.vue')
+      }
+    ]
+  },
+  {
+    path: '/tabs-secondary/',
+    component: () => import('@/views/TabsSecondary.vue'),
+    children: [
+      {
+        path: '',
+        redirect: '/tabs-secondary/tab1'
+      },
+      {
+        path: 'tab1',
+        component: () => import('@/views/Tab1Secondary.vue')
+      },
+      {
+        path: 'tab2',
+        component: () => import('@/views/Tab2Secondary.vue')
+      },
+      {
+        path: 'tab3',
+        component: () => import('@/views/Tab3Secondary.vue')
       }
     ]
   }
@@ -87,6 +133,6 @@ const routes: Array<RouteRecordRaw> = [
 const router = createRouter({
   history: createWebHistory(process.env.BASE_URL),
   routes
-})
+});
 
 export default router